+++ src/components/Gnb/HeaderMypage.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import SkeletonUser from "./skeletonUser";
-import { MemberType, UserType } from "./type";
+import SkeletonUser from "../../shared/skeletonUser";
+import { MemberType, UserType } from "../../types/users";
 import { getMembers } from "@/api/members";
 import { getUserInfo } from "@/api/user";
 import RandomProfile from "../table/member/RandomProfile";
@@ -105,7 +105,7 @@ const HeaderMyPage = () => {
                 {members.slice(0, 4).map((member) => (
                   <img
                     key={member.id}
-                    src={member.profileImageUrl || "/svgs/dummy-icon.png"}
+                    src={member.profileImageUrl}
                     alt={member.nickname}
                     className="w-[34px] h-[34px] md:w-[38px] md:h-[38px] rounded-full border-[2px] border-white"
                   />
