--- src/components/Gnb/type.ts
@@ -1,10 +1,0 @@
-export interface MemberType {
-  id: number;
-  profileImageUrl: string;
-  nickname: string;
-}
-
-export interface UserType {
-  profileImageUrl: string;
-  nickname: string;
-}
