+++ src/types/users.ts
@@ -1,0 +1,8 @@
+export interface UserType {
+  profileImageUrl: string;
+  nickname: string;
+}
+
+export interface MemberType extends UserType {
+  id: number;
+}
